moved host object to variorum.c
