+++ src/powmon/common.c
@@ -81,7 +81,7 @@ void parse_json_obj(char *s, int num_sockets, char *hostname)
     for (i = 0; i < num_sockets; i++)
     {
         /* extract socket object from node object with "Socket_#" */
-        snprintf(socketID, 20, "Socket_%d", i);
+        snprintf(socketID, 20, "socket_%d", i);
         json_t *socket_obj = json_object_get(node_obj, socketID);
         if (socket_obj == NULL)
         {
