+++ src/variorum/ARM/neoverse_N1_power_features.c
@@ -244,26 +244,9 @@ int arm_cpu_neoverse_n1_cap_socket_frequency(int socketid, int new_freq)
 
 int arm_cpu_neoverse_n1_json_get_power_data(json_t *get_power_obj)
 {
-    char hostname[1024];
-    struct timeval tv;
-    uint64_t ts;
-
     uint64_t cpu_power_val;
     uint64_t io_power_val;
     int i;
-
-    gethostname(hostname, 1024);
-    gettimeofday(&tv, NULL);
-    ts = tv.tv_sec * (uint64_t)1000000 + tv.tv_usec;
-
-    json_t *node_obj = json_object_get(get_power_obj, hostname);
-    if (node_obj == NULL)
-    {
-        node_obj = json_object();
-        json_object_set_new(get_power_obj, hostname, node_obj);
-    }
-
-    json_object_set_new(node_obj, "timestamp", json_integer(ts));
 
     /* Read power data from hwmon interfaces, similar to the get_power_data()
        function, defined previously. */
@@ -305,13 +288,13 @@ int arm_cpu_neoverse_n1_json_get_power_data(json_t *get_power_obj)
     for (i = 0; i < (int)m_num_package; i++)
     {
         char socketID[16];
-        snprintf(socketID, 16, "Socket_%d", i);
+        snprintf(socketID, 16, "socket_%d", i);
 
-        json_t *socket_obj = json_object_get(node_obj, socketID);
+        json_t *socket_obj = json_object_get(get_power_obj, socketID);
         if (socket_obj == NULL)
         {
             socket_obj = json_object();
-            json_object_set_new(node_obj, socketID, socket_obj);
+            json_object_set_new(get_power_obj, socketID, socket_obj);
         }
 
         json_t *cpu_obj = json_object();
@@ -331,7 +314,7 @@ int arm_cpu_neoverse_n1_json_get_power_data(json_t *get_power_obj)
        Variorum converts power into watts before reporting. Socket 0 is big,
        and Socket 1 is little. */
 
-    json_t *socket_0_obj = json_object_get(node_obj, "Socket_0");
+    json_t *socket_0_obj = json_object_get(node_obj, "socket_0");
     json_t *socket_0_cpu_obj = json_object_get(socket_0_obj, "CPU");
 
     json_t *socket_0_gpu_obj = json_object();
